backend/requirements.txt
fastapi==0.124.4
python-dotenv==1.2.1
uvicorn==0.38.0
dotenv==0.9.9
greenlet==3.3.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
passlib==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.3.0
dnspython==2.8.0
geoalchemy2==0.18.1
minio==7.2.20
python-magic==0.4.27
python-multipart==0.0.22
Pillow==12.1.1
redis==5.0.1